refactor: code cleanup & adhere to current file structure
