+++ src/mobile/feature-counter/index.ts
@@ -1,0 +1,2 @@
+export * from "./components/CounterPage";
+export * from "./components/CounterNoData";
